Login page; promise in sunc; organize codes
