+++ mainPage.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 
 <head>
-    <title>Zoom WebSDK</title>
+    <title>Gaze Learning</title>
     <meta charset="utf-8" />
     <link type="text/css" rel="stylesheet" href="https://source.zoom.us/1.8.1/css/bootstrap.css" />
     <link type="text/css" rel="stylesheet" href="https://source.zoom.us/1.8.1/css/react-select.css" />
@@ -9,8 +9,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no">
 
     <script src="https://api.gazerecorder.com/GazeCloudAPI.js"></script>
-    <script src="./webgazer.js"></script>
-    <script src="./heatmap.js"></script>
+    <script src="js/webgazer.js"></script>
+    <script src="js/heatmap.js"></script>
 
     <style type="text/css">
         body {
@@ -83,7 +83,7 @@
         <!-- <img src='regression.jpg'> -->
     </div>
     
-    <script src="./client_original.js"></script>
+    <script src="js/client.js"></script>
 
     <script src="Zoom/tool.js"></script>
     <script src="Zoom/embed.js"></script>
