+++ startMeeting.html
@@ -1,7 +1,9 @@
 <!DOCTYPE html>
+<!-- This page is embedded into mainPage.html via iframe, it configs the meeting. -->
+<!-- This page will then jump to meeting.html -->
 
 <head>
-    <title>Zoom WebSDK</title>
+    <!-- <title>Zoom WebSDK</title> -->
     <meta charset="utf-8" />
     <link type="text/css" rel="stylesheet" href="https://source.zoom.us/1.8.1/css/bootstrap.css" />
     <link type="text/css" rel="stylesheet" href="https://source.zoom.us/1.8.1/css/react-select.css" />
